Update Import-VcConfigMgrApplication.ps1
I kept getting an error when running the script:

Get-Item : Cannot find path '\\<myserver>\<myshare>' because it does not exist.
At C:\Program Files\WindowsPowerShell\Modules\VcRedist\2.0.132\Public\Import-VcConfigMgrApplication.ps1:183 char:68

I isolated it to when the Deployment Type was being created in SCCM (version 1810). It's due to set-location on line 177 and get-item on line 183. Once I updated the code everything worked as expected.

Source: https://stackoverflow.com/questions/23574653/running-get-childitem-on-unc-path-works-in-powershell-but-not-in-powershell-run